+++ src/chainmail/service/__init__.py
@@ -2,9 +2,10 @@
 
 from .client import GovernorClient, GovernorClientError
 from .protocol import ProtocolError, proposal_from_dict, result_to_dict
-from .server import UnixSocketGovernorServer, main
+from .server import CallerIdentity, UnixSocketGovernorServer, main
 
 __all__ = [
+    "CallerIdentity",
     "GovernorClient",
     "GovernorClientError",
     "UnixSocketGovernorServer",